requirements.txt
jinja2==3.1.2
pandas==2.1.3
pip==23.3.1
python-dotenv[cli]==1.0.0
python-gitlab==4.1.1
